requirement.txt
python-dotenv==1.0.0
openai==0.27.8
langchain==0.0.268